requirements.txt
scikit-learn
scipy
numpy
python-dotenv
azure-cognitiveservices-vision-computervision
msrest